Fix notes getting assigned to unused outputs
If 4 outputs are used, but 5 simultaneous notes arrived, the fifth note
would be output on the 5th CV/Gate pair and get stuff on that pair the
next time it was played. The check to find if a note is already being
played wasn't actually checking if the note was on, only if it was
present in the list. This wasn't possible before the previous commit.