Complete rework for new midi modes and calibration
Added a bit of new framework that allows modes to be defined that share
common stuff but implement different behaviors. The original
16 voice polyphonic mode is preserved as the first mode.

Config mode is for changing the midi channel. It was previously
separate from the other modes because of some special switch handling
behaviors. That's been reworked so that the config mode can be like
other modes accept that when config mode is entered, the saved mode in
EEPROM isn't changed. The saved mode is restored when config mode is
exited. Handling of the switch has been improved a bit and now only
acts when the switch is released (accept for the panic function).

A mode for using all 32 outputs as if they were gates was added. this
mode is meant for something like drum triggers. It responds to note
on/off 0-31 and each note maps to gate or cv.

The polyphonic harmonic mode is a work in progress mode that uses 4 CVs
to play a note and it's first three harmonics. It currently doesn't
account for the calibration correctly in the second harmonic.

Test functionality was converted to a mode as well. It displays a
pattern on the gates and ramps the CV outputs through the full range.

A per note calibration has been added that corrects the CV at each note
for each CV output. The calibration requires additional scripting in
another repository, refer to the README.